Gen_Ai/task 3/requirements.txt
langchain
langchain-groq
python-dotenv
langsmith
streamlit
pypdf
pydantic